ci-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile ci-requirements.in
#
agate==1.7.1
    # via dbt-core
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via gcsfs
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
babel==2.17.0
    # via agate
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via requests
cffi==1.17.1
    # via dbt-core
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
colorama==0.4.6
    # via
    #   click
    #   dbt-core
db-dtypes==1.4.3
    # via
    #   google-cloud-bigquery
    #   pandas-gbq
dbt-bigquery==1.7.9
    # via -r C:\Users\George.Singer\Documents\GitHub\Arial\dbt_requirements.in
dbt-core==1.7.19
    # via dbt-bigquery
dbt-extractor==0.5.1
    # via dbt-core
dbt-semantic-interfaces==0.4.4
    # via dbt-core
decorator==5.2.1
    # via gcsfs
flake8==7.3.0
    # via -r C:\Users\George.Singer\Documents\GitHub\Arial\dbt_requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.7.0
    # via gcsfs
gcsfs==2025.7.0
    # via -r C:\Users\George.Singer\Documents\GitHub\Arial\requirements.in
google-api-core[grpc]==2.25.1
    # via
    #   dbt-bigquery
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-dataproc
    #   google-cloud-storage
    #   pandas-gbq
google-auth==2.40.3
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-dataproc
    #   google-cloud-storage
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==1.2.2
    # via
    #   gcsfs
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery[pandas]==3.35.1
    # via
    #   -r C:\Users\George.Singer\Documents\GitHub\Arial\requirements.in
    #   dbt-bigquery
    #   pandas-gbq
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-dataproc==5.21.0
    # via dbt-bigquery
google-cloud-storage==2.19.0
    # via
    #   dbt-bigquery
    #   gcsfs
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-dataproc
grpcio==1.74.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
idna==3.10
    # via
    #   dbt-core
    #   requests
    #   yarl
importlib-metadata==6.11.0
    # via dbt-semantic-interfaces
isodate==0.6.1
    # via
    #   agate
    #   dbt-core
jinja2==3.1.6
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
joblib==1.5.1
    # via scikit-learn
jsonschema==4.25.0
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
jsonschema-specifications==2025.4.1
    # via jsonschema
leather==0.4.0
    # via agate
logbook==1.5.3
    # via dbt-core
markupsafe==3.0.2
    # via jinja2
mashumaro[msgpack]==3.14
    # via dbt-core
mccabe==0.7.0
    # via flake8
minimal-snowplow-tracker==0.0.2
    # via dbt-core
more-itertools==10.7.0
    # via dbt-semantic-interfaces
msgpack==1.1.1
    # via mashumaro
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via dbt-core
numpy==2.2.6
    # via
    #   db-dtypes
    #   pandas
    #   pandas-gbq
    #   scikit-learn
    #   scipy
    #   xgboost
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via
    #   db-dtypes
    #   dbt-core
    #   google-cloud-bigquery
    #   pandas-gbq
pandas==2.3.1
    # via
    #   -r C:\Users\George.Singer\Documents\GitHub\Arial\requirements.in
    #   db-dtypes
    #   google-cloud-bigquery
    #   pandas-gbq
pandas-gbq==0.29.2
    # via google-cloud-bigquery
parsedatetime==2.6
    # via agate
pathspec==0.11.2
    # via dbt-core
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-dataproc
protobuf==4.25.8
    # via
    #   dbt-core
    #   google-api-core
    #   google-cloud-dataproc
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==21.0.0
    # via
    #   db-dtypes
    #   google-cloud-bigquery
    #   pandas-gbq
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycodestyle==2.14.0
    # via flake8
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via dbt-semantic-interfaces
pydantic-core==2.33.2
    # via pydantic
pydata-google-auth==1.9.1
    # via pandas-gbq
pyflakes==3.4.0
    # via flake8
python-dateutil==2.9.0.post0
    # via
    #   dbt-semantic-interfaces
    #   google-cloud-bigquery
    #   pandas
python-dotenv==1.1.1
    # via -r C:\Users\George.Singer\Documents\GitHub\Arial\requirements.in
python-slugify==8.0.4
    # via agate
pytimeparse==1.1.8
    # via agate
pytz==2025.2
    # via
    #   dbt-core
    #   pandas
pyyaml==6.0.2
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   -r C:\Users\George.Singer\Documents\GitHub\Arial\requirements.in
    #   dbt-core
    #   gcsfs
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   minimal-snowplow-tracker
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
scikit-learn==1.7.1
    # via -r C:\Users\George.Singer\Documents\GitHub\Arial\requirements.in
scipy==1.15.3
    # via
    #   scikit-learn
    #   xgboost
six==1.17.0
    # via
    #   isodate
    #   minimal-snowplow-tracker
    #   python-dateutil
sqlparse==0.5.3
    # via dbt-core
text-unidecode==1.3
    # via python-slugify
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   dbt-core
    #   dbt-semantic-interfaces
    #   mashumaro
    #   multidict
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==1.26.20
    # via
    #   dbt-core
    #   requests
xgboost==3.0.2
    # via -r C:\Users\George.Singer\Documents\GitHub\Arial\requirements.in
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
